[thpatch] [UI] Hide list view and pagemode
What is this, even?

On the one hand, we have one "list" mode with a completely different
design compared to the two page and review modes. Since 2 is greater
than 1, let's focus our redesign on the page and review modes…

… but as we saw earlier, pagemode itself is just a subset of review
mode. So, let's also throw out that one, and exclusively focus the
upcoming redesign commits on review mode.

Change-Id: I6be156f14007945910339dcea59def6e343a1593