[thpatch] [UI] Move the edit-mark icon to the left of tux-proofread-status
Since the background color already communicates whether a message is
untranslated, translated, fuzzy, or proofread, the status column on its
own is pretty redundant. I do agree that the edit-mark icon does serve
a useful role though, in communicating that clicking the hovered message
will edit it.
But hey, the previous tux-proofread-status icons can fulfill the same
role if we only show them when hovering over the message! And using
the edit-mark only for already translated messages makes perfect sense.
It may be called *edit*-mark for a reason!

Thankfully, edit-mark.svg is exactly the same size as the other icons…
as if upstream actually had this superior idea before. 🤔

Change-Id: Iff688d69a192c7e455ebeae17442bc15945c0efa