[thpatch] [UI] Remove stickyness of the message header
Um, with the bigger batch sizes we prefer to use, changing the filter
is a rather costly operation. With both "infinite" scroll and the
background colors, it really doesn't need to be a permanent part of the
UI in our redesign… especially if it makes translation unit anchors work
again. No need for CSS `scroll-padding`, and the accompanied worries
about Safari users, who would require their own workaround for the
workaround.

Change-Id: I3ecb325f381b0a6069b5ff5d717dc41d2abfd779