[thpatch] Remove the priority language section from Special:PageTranslation
We'll be running a fork of this extension for the foreseeable future
anyway. It's more robust to remove this section from PHP rather than
hiding it via a CSS rule that can easily get outdated whenever upstream
adds new sections to the page (which did in fact happen between 2014
and now). In PHP, we'd at least get a merge conflict in that case.

Change-Id: I85c79d98462be43479611c417ab5763e1ef107b4